requirements.txt
# Core
openai>=1.30.0
langgraph>=0.1.0
langchain-core>=0.2.0

# API & Server
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Data validation
pydantic>=2.7.0
python-dotenv>=1.0.0

# Utilities
httpx>=0.27.0
aiofiles>=23.2.1
